lost_found_app/requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
cloudinary>=1.36.0
